refactor: split schema.ts table SQL by domain (#585)
Pure refactor — no logic changes. The 966-line `schema.ts` now imports
per-domain table SQL maps from `schema/`:

- schema/activities.ts (269)   — activities, activity_type_definitions
                                  (with 80+-row built-in seed), deduction
                                  rules, legacy tag tables
- schema/meals.ts (209)        — meals, meal_log_completed, food_items,
                                  meal_food_items
- schema/system.ts (154)       — oauth, mcp_sessions, sync queue + state,
                                  user_settings, audit_log, notes, goals,
                                  uploaded_icons
- schema/locations.ts (81)     — detected_locations, locations,
                                  named_locations, places
- schema/reports.ts (64)       — lab_results, reports, report_entries
- schema/metrics.ts (58)       — time_series, custom_metrics, raw_records
- schema/productivity.ts (50)  — productivity, screentime_categories

`schema.ts` (138 lines) retains api-spec re-exports, SCHEMA_VERSION, and
the dependency-respecting `tableCreationOrder`. No consumer imports change.
