+++ apps/backend/src/schema.ts
@@ -1,9 +1,20 @@
 /**
  * Database schema definitions for Aurboda.
  *
- * This module contains all table creation SQL and schema-related utilities.
+ * Per-domain table SQL lives in `schema/`; this file assembles the full
+ * `createTableStatements` map and the dependency-respecting
+ * `tableCreationOrder` array consumed by db/connection.ts and migrate.ts.
+ *
  * See docs/data-storage.md for design decisions and data flow documentation.
  */
+
+import { activitiesTables } from './schema/activities.ts'
+import { locationsTables } from './schema/locations.ts'
+import { mealsTables } from './schema/meals.ts'
+import { metricsTables } from './schema/metrics.ts'
+import { productivityTables } from './schema/productivity.ts'
+import { reportsTables } from './schema/reports.ts'
+import { systemTables } from './schema/system.ts'
 
 // Re-export common types from shared api-spec package
 export {
@@ -33,866 +44,27 @@ export const SCHEMA_VERSION = 1
 
 /**
  * All table creation statements in dependency order.
- * Each table is created with proper indexes for query performance.
+ * Assembled from per-domain SQL modules under `schema/`.
  */
 export const createTableStatements: Record<string, string> = {
-  // Unified activities table (covers all time-ranged events: sleep, exercise, tags, calendar, etc.)
-  activities: `
-    CREATE TABLE IF NOT EXISTS activities (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      source          VARCHAR(50) NOT NULL,
-      external_id     VARCHAR(255),
-      activity_type   VARCHAR(100) NOT NULL REFERENCES activity_type_definitions(name) ON UPDATE CASCADE,
-      start_time      TIMESTAMPTZ NOT NULL,
-      end_time        TIMESTAMPTZ,
-      title           VARCHAR(255),
-      notes           TEXT,
-      data            JSONB,
-      deleted_at      TIMESTAMPTZ,
-      superseded_by   UUID REFERENCES activities(id) ON DELETE SET NULL
-    )
-  `,
-
-  activities_indexes: `
-    CREATE UNIQUE INDEX IF NOT EXISTS idx_activities_ext_id ON activities (source, external_id) WHERE external_id IS NOT NULL;
-    CREATE UNIQUE INDEX IF NOT EXISTS idx_activities_type_time ON activities (source, activity_type, start_time) WHERE external_id IS NULL;
-    CREATE INDEX IF NOT EXISTS idx_activities_time_range ON activities (start_time, end_time);
-    CREATE INDEX IF NOT EXISTS idx_activities_not_deleted ON activities (activity_type, start_time DESC) WHERE deleted_at IS NULL;
-    CREATE INDEX IF NOT EXISTS idx_activities_not_superseded ON activities (activity_type, start_time DESC) WHERE deleted_at IS NULL AND superseded_by IS NULL
-  `,
-
-  // Detected locations (clusters detected from GPS data with geocoded addresses)
-  detected_locations: `
-    CREATE TABLE IF NOT EXISTS detected_locations (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      location        GEOGRAPHY(POINT, 4326) NOT NULL,
-      radius          INTEGER NOT NULL DEFAULT 200,
-      total_minutes   INTEGER NOT NULL DEFAULT 0,
-      visit_count     INTEGER NOT NULL DEFAULT 0,
-      first_visit     TIMESTAMPTZ NOT NULL,
-      last_visit      TIMESTAMPTZ NOT NULL,
-      address         TEXT,
-      geocode_status  VARCHAR(20) DEFAULT 'pending',
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      updated_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-
-  detected_locations_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_detected_locations_geo
-      ON detected_locations USING GIST (location);
-    CREATE INDEX IF NOT EXISTS idx_detected_locations_geocode_status
-      ON detected_locations (geocode_status) WHERE geocode_status = 'pending'
-  `,
-
-  // Lab results / blood work (legacy flat table — superseded by reports + report_entries)
-  lab_results: `
-    CREATE TABLE IF NOT EXISTS lab_results (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      test_date       DATE NOT NULL,
-      test_name       VARCHAR(100) NOT NULL,
-      test_category   VARCHAR(50),
-      value           DOUBLE PRECISION NOT NULL,
-      unit            VARCHAR(30) NOT NULL,
-      reference_low   DOUBLE PRECISION,
-      reference_high  DOUBLE PRECISION,
-      flag            VARCHAR(10),
-      lab_name        VARCHAR(100),
-      notes           TEXT,
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-
-  lab_results_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_lab_results_date ON lab_results (test_date DESC);
-    CREATE INDEX IF NOT EXISTS idx_lab_results_test ON lab_results (test_name, test_date DESC);
-    CREATE INDEX IF NOT EXISTS idx_lab_results_category ON lab_results (test_category, test_date DESC)
-  `,
-
-  // GPS location data with PostGIS support
-  locations: `
-    CREATE TABLE IF NOT EXISTS locations (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      source          VARCHAR(50) NOT NULL DEFAULT 'owntracks',
-      time            TIMESTAMPTZ NOT NULL,
-      location        GEOGRAPHY(POINT, 4326) NOT NULL,
-      accuracy        DOUBLE PRECISION,
-      altitude        DOUBLE PRECISION,
-      velocity        DOUBLE PRECISION,
-      regions         VARCHAR[] DEFAULT '{}',
-      deleted_at      TIMESTAMPTZ,
-      CONSTRAINT unique_location UNIQUE (source, time)
-    )
-  `,
-  locations_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_locations_time ON locations (time DESC);
-    CREATE INDEX IF NOT EXISTS idx_locations_geo ON locations USING GIST (location)
-  `,
-
-  // MCP session persistence for surviving backend restarts
-
-  mcp_sessions: `
-    CREATE TABLE IF NOT EXISTS mcp_sessions (
-      session_id      UUID PRIMARY KEY,
-      username        VARCHAR(255) NOT NULL,
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      last_activity   TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-
-  mcp_sessions_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_mcp_sessions_username ON mcp_sessions (username);
-    CREATE INDEX IF NOT EXISTS idx_mcp_sessions_last_activity ON mcp_sessions (last_activity)
-  `,
-
-  // Meal/nutrition data
-  meals: `
-    CREATE TABLE IF NOT EXISTS meals (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      source          VARCHAR(50) NOT NULL DEFAULT 'manual',
-      meal_type       VARCHAR(50),
-      name            VARCHAR(255),
-      time            TIMESTAMPTZ NOT NULL,
-      calories        DOUBLE PRECISION,
-      protein         DOUBLE PRECISION,
-      carbs           DOUBLE PRECISION,
-      fat             DOUBLE PRECISION,
-      fiber           DOUBLE PRECISION,
-      food_items      JSONB,
-      micros          JSONB,
-      notes           TEXT,
-      sensitivities   TEXT[],
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-
-  // Meal log completion flag per day
-  meal_log_completed: `
-    CREATE TABLE IF NOT EXISTS meal_log_completed (
-      date            DATE PRIMARY KEY,
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-
-  // Canonical food item library (normalized from JSONB)
-  food_items: `
-    CREATE TABLE IF NOT EXISTS food_items (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      name            VARCHAR(255) NOT NULL,
-      name_lower      VARCHAR(255) NOT NULL,
-      source          VARCHAR(50) NOT NULL DEFAULT 'manual',
-      default_quantity DOUBLE PRECISION,
-      default_unit    VARCHAR(100),
-      calories        DOUBLE PRECISION,
-      protein         DOUBLE PRECISION,
-      carbs           DOUBLE PRECISION,
-      fat             DOUBLE PRECISION,
-      fiber           DOUBLE PRECISION,
-      alcohol         DOUBLE PRECISION,
-      caffeine        DOUBLE PRECISION,
-      water           DOUBLE PRECISION,
-      net_carbs       DOUBLE PRECISION,
-      starch          DOUBLE PRECISION,
-      sugars          DOUBLE PRECISION,
-      added_sugars    DOUBLE PRECISION,
-      cholesterol     DOUBLE PRECISION,
-      saturated_fat   DOUBLE PRECISION,
-      monounsaturated_fat DOUBLE PRECISION,
-      polyunsaturated_fat DOUBLE PRECISION,
-      trans_fat       DOUBLE PRECISION,
-      omega_3         DOUBLE PRECISION,
-      omega_6         DOUBLE PRECISION,
-      ala             DOUBLE PRECISION,
-      dha             DOUBLE PRECISION,
-      epa             DOUBLE PRECISION,
-      dpa             DOUBLE PRECISION,
-      aa              DOUBLE PRECISION,
-      la              DOUBLE PRECISION,
-      vitamin_a       DOUBLE PRECISION,
-      retinol         DOUBLE PRECISION,
-      beta_carotene   DOUBLE PRECISION,
-      vitamin_c       DOUBLE PRECISION,
-      vitamin_d       DOUBLE PRECISION,
-      vitamin_e       DOUBLE PRECISION,
-      vitamin_k       DOUBLE PRECISION,
-      b1_thiamine     DOUBLE PRECISION,
-      b2_riboflavin   DOUBLE PRECISION,
-      b3_niacin       DOUBLE PRECISION,
-      b5_pantothenic_acid DOUBLE PRECISION,
-      b6_pyridoxine   DOUBLE PRECISION,
-      b12_cobalamin   DOUBLE PRECISION,
-      folate          DOUBLE PRECISION,
-      calcium         DOUBLE PRECISION,
-      copper          DOUBLE PRECISION,
-      iron            DOUBLE PRECISION,
-      magnesium       DOUBLE PRECISION,
-      manganese       DOUBLE PRECISION,
-      phosphorus      DOUBLE PRECISION,
-      potassium       DOUBLE PRECISION,
-      selenium        DOUBLE PRECISION,
-      sodium          DOUBLE PRECISION,
-      zinc            DOUBLE PRECISION,
-      iodine          DOUBLE PRECISION,
-      cystine         DOUBLE PRECISION,
-      histidine       DOUBLE PRECISION,
-      isoleucine      DOUBLE PRECISION,
-      leucine         DOUBLE PRECISION,
-      lysine          DOUBLE PRECISION,
-      methionine      DOUBLE PRECISION,
-      phenylalanine   DOUBLE PRECISION,
-      threonine       DOUBLE PRECISION,
-      tryptophan      DOUBLE PRECISION,
-      tyrosine        DOUBLE PRECISION,
-      valine          DOUBLE PRECISION,
-      oxalate         DOUBLE PRECISION,
-      phytate         DOUBLE PRECISION,
-      ash             DOUBLE PRECISION,
-      salt            DOUBLE PRECISION,
-      icon            TEXT,
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      updated_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      CONSTRAINT unique_food_item_name UNIQUE (name_lower)
-    )
-  `,
-
-  food_items_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_food_items_name_lower ON food_items (name_lower)
-  `,
-
-  // Junction: meals <-> food items (snapshot of nutrients at insertion time)
-  meal_food_items: `
-    CREATE TABLE IF NOT EXISTS meal_food_items (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      meal_id         UUID NOT NULL REFERENCES meals(id) ON DELETE CASCADE,
-      food_item_id    UUID NOT NULL REFERENCES food_items(id),
-      quantity        DOUBLE PRECISION,
-      unit            VARCHAR(100),
-      sort_order      INTEGER NOT NULL DEFAULT 0,
-      calories        DOUBLE PRECISION,
-      protein         DOUBLE PRECISION,
-      carbs           DOUBLE PRECISION,
-      fat             DOUBLE PRECISION,
-      fiber           DOUBLE PRECISION,
-      alcohol         DOUBLE PRECISION,
-      caffeine        DOUBLE PRECISION,
-      water           DOUBLE PRECISION,
-      net_carbs       DOUBLE PRECISION,
-      starch          DOUBLE PRECISION,
-      sugars          DOUBLE PRECISION,
-      added_sugars    DOUBLE PRECISION,
-      cholesterol     DOUBLE PRECISION,
-      saturated_fat   DOUBLE PRECISION,
-      monounsaturated_fat DOUBLE PRECISION,
-      polyunsaturated_fat DOUBLE PRECISION,
-      trans_fat       DOUBLE PRECISION,
-      omega_3         DOUBLE PRECISION,
-      omega_6         DOUBLE PRECISION,
-      ala             DOUBLE PRECISION,
-      dha             DOUBLE PRECISION,
-      epa             DOUBLE PRECISION,
-      dpa             DOUBLE PRECISION,
-      aa              DOUBLE PRECISION,
-      la              DOUBLE PRECISION,
-      vitamin_a       DOUBLE PRECISION,
-      retinol         DOUBLE PRECISION,
-      beta_carotene   DOUBLE PRECISION,
-      vitamin_c       DOUBLE PRECISION,
-      vitamin_d       DOUBLE PRECISION,
-      vitamin_e       DOUBLE PRECISION,
-      vitamin_k       DOUBLE PRECISION,
-      b1_thiamine     DOUBLE PRECISION,
-      b2_riboflavin   DOUBLE PRECISION,
-      b3_niacin       DOUBLE PRECISION,
-      b5_pantothenic_acid DOUBLE PRECISION,
-      b6_pyridoxine   DOUBLE PRECISION,
-      b12_cobalamin   DOUBLE PRECISION,
-      folate          DOUBLE PRECISION,
-      calcium         DOUBLE PRECISION,
-      copper          DOUBLE PRECISION,
-      iron            DOUBLE PRECISION,
-      magnesium       DOUBLE PRECISION,
-      manganese       DOUBLE PRECISION,
-      phosphorus      DOUBLE PRECISION,
-      potassium       DOUBLE PRECISION,
-      selenium        DOUBLE PRECISION,
-      sodium          DOUBLE PRECISION,
-      zinc            DOUBLE PRECISION,
-      iodine          DOUBLE PRECISION,
-      cystine         DOUBLE PRECISION,
-      histidine       DOUBLE PRECISION,
-      isoleucine      DOUBLE PRECISION,
-      leucine         DOUBLE PRECISION,
-      lysine          DOUBLE PRECISION,
-      methionine      DOUBLE PRECISION,
-      phenylalanine   DOUBLE PRECISION,
-      threonine       DOUBLE PRECISION,
-      tryptophan      DOUBLE PRECISION,
-      tyrosine        DOUBLE PRECISION,
-      valine          DOUBLE PRECISION,
-      oxalate         DOUBLE PRECISION,
-      phytate         DOUBLE PRECISION,
-      ash             DOUBLE PRECISION,
-      salt            DOUBLE PRECISION
-    )
-  `,
-
-  meal_food_items_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_meal_food_items_meal ON meal_food_items (meal_id);
-    CREATE INDEX IF NOT EXISTS idx_meal_food_items_food ON meal_food_items (food_item_id)
-  `,
-
-  meals_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_meals_time ON meals (time DESC);
-    CREATE INDEX IF NOT EXISTS idx_meals_type_time ON meals (meal_type, time DESC);
-    CREATE INDEX IF NOT EXISTS idx_meals_source ON meals (source, time DESC)
-  `,
-
-  // User-defined named locations (detected and named via Aurboda)
-  named_locations: `
-    CREATE TABLE IF NOT EXISTS named_locations (
-      id                    UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      name                  VARCHAR(255) NOT NULL,
-      location              GEOGRAPHY(POINT, 4326) NOT NULL,
-      radius                INTEGER NOT NULL DEFAULT 200,
-      auto_create_activity  BOOLEAN NOT NULL DEFAULT false,
-      created_at            TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      updated_at            TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-  named_locations_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_named_locations_geo ON named_locations USING GIST (location)
-  `,
-
-  // Notes/comments on any entity (polymorphic reference)
-  notes: `
-    CREATE TABLE IF NOT EXISTS notes (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      entity_type     VARCHAR(50) NOT NULL,
-      entity_id       TEXT NOT NULL,
-      content         TEXT NOT NULL,
-      source          VARCHAR(50),
-      start_time      TIMESTAMPTZ,
-      end_time        TIMESTAMPTZ,
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      updated_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-
-  notes_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_notes_entity ON notes (entity_type, entity_id);
-    CREATE INDEX IF NOT EXISTS idx_notes_created ON notes (created_at DESC);
-    CREATE INDEX IF NOT EXISTS idx_notes_time ON notes (start_time, end_time)
-  `,
-
-  // OAuth tokens for third-party APIs
-  oauth_tokens: `
-    CREATE TABLE IF NOT EXISTS oauth_tokens (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      provider        VARCHAR(50) NOT NULL,
-      access_token    TEXT NOT NULL,
-      refresh_token   TEXT,
-      expires_at      TIMESTAMPTZ,
-      scopes          VARCHAR[],
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      updated_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      CONSTRAINT unique_provider UNIQUE (provider)
-    )
-  `,
-
-  // Outbound sync queue - tracks changes to push to Health Connect
-  outbound_sync_queue: `
-    CREATE TABLE IF NOT EXISTS outbound_sync_queue (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      entity_type     VARCHAR(50) NOT NULL,
-      entity_id       VARCHAR(255) NOT NULL,
-      operation       VARCHAR(20) NOT NULL,
-      hc_record_type  VARCHAR(100) NOT NULL,
-      payload         JSONB NOT NULL,
-      hc_record_id    VARCHAR(255),
-      status          VARCHAR(20) NOT NULL DEFAULT 'pending',
-      fail_count      INT NOT NULL DEFAULT 0,
-      fail_reason     TEXT,
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      synced_at       TIMESTAMPTZ
-    )
-  `,
-  outbound_sync_queue_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_outbound_sync_queue_status ON outbound_sync_queue (status) WHERE status = 'pending';
-    CREATE INDEX IF NOT EXISTS idx_outbound_sync_queue_entity ON outbound_sync_queue (entity_type, entity_id)
-  `,
-
-  // Named places / geofences
-
-  places: `
-    CREATE TABLE IF NOT EXISTS places (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      source          VARCHAR(50) NOT NULL DEFAULT 'owntracks',
-      external_id     VARCHAR(255),
-      name            VARCHAR(255) NOT NULL,
-      location        GEOGRAPHY(POINT, 4326) NOT NULL,
-      radius          INTEGER NOT NULL,
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      CONSTRAINT unique_place UNIQUE (source, external_id)
-    )
-  `,
-  places_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_places_geo ON places USING GIST (location)
-  `,
-
-  // Productivity data (RescueTime, ActivityWatch, etc.)
-
-  productivity: `
-    CREATE TABLE IF NOT EXISTS productivity (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      source          VARCHAR(50) NOT NULL DEFAULT 'rescuetime',
-      start_time      TIMESTAMPTZ NOT NULL,
-      end_time        TIMESTAMPTZ NOT NULL,
-      activity        VARCHAR(255) NOT NULL,
-      title           TEXT,
-      category        VARCHAR(100),
-      productivity    SMALLINT,
-      duration_sec    INTEGER NOT NULL,
-      is_mobile       BOOLEAN DEFAULT FALSE,
-      device_name     VARCHAR(100) NOT NULL DEFAULT '',
-      resolved_category TEXT[],
-      deleted_at      TIMESTAMPTZ,
-      CONSTRAINT unique_productivity UNIQUE (source, start_time, activity, device_name)
-    )
-  `,
-  productivity_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_productivity_time ON productivity (start_time DESC);
-    CREATE INDEX IF NOT EXISTS idx_productivity_category ON productivity (category, start_time DESC);
-    CREATE INDEX IF NOT EXISTS idx_productivity_not_deleted ON productivity (start_time DESC) WHERE deleted_at IS NULL;
-    CREATE INDEX IF NOT EXISTS idx_productivity_resolved_category ON productivity (resolved_category) WHERE resolved_category IS NOT NULL
-  `,
-
-  // Raw data sink - stores all incoming data in original form
-
-  raw_records: `
-    CREATE TABLE IF NOT EXISTS raw_records (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      source          VARCHAR(50) NOT NULL,
-      record_type     VARCHAR(100) NOT NULL,
-      external_id     VARCHAR(255),
-      recorded_at     TIMESTAMPTZ NOT NULL,
-      received_at     TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      data            JSONB NOT NULL,
-      CONSTRAINT unique_source_record UNIQUE (source, record_type, external_id)
-    )
-  `,
-  raw_records_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_raw_records_source_time ON raw_records (source, recorded_at);
-    CREATE INDEX IF NOT EXISTS idx_raw_records_type_time ON raw_records (record_type, recorded_at);
-    CREATE INDEX IF NOT EXISTS idx_raw_records_data ON raw_records USING GIN (data)
-  `,
-
-  report_entries: `
-    CREATE TABLE IF NOT EXISTS report_entries (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      report_id       UUID NOT NULL REFERENCES reports(id) ON DELETE CASCADE,
-      metric          VARCHAR(100) NOT NULL,
-      value           DOUBLE PRECISION,
-      unit            VARCHAR(30),
-      method          VARCHAR(50),
-      confidence      VARCHAR(20),
-      reference_low   DOUBLE PRECISION,
-      reference_high  DOUBLE PRECISION,
-      flag            VARCHAR(20)
-    )
-  `,
-  report_entries_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_report_entries_report ON report_entries (report_id);
-    CREATE INDEX IF NOT EXISTS idx_report_entries_metric ON report_entries (metric)
-  `,
-
-  // Structured lab reports (InBody, blood panels, hair mineral analysis, etc.)
-  reports: `
-    CREATE TABLE IF NOT EXISTS reports (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      report_type     VARCHAR(100) NOT NULL,
-      report_date     TIMESTAMPTZ NOT NULL,
-      location        VARCHAR(255),
-      notes           TEXT,
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-  reports_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_reports_date ON reports (report_date DESC);
-    CREATE INDEX IF NOT EXISTS idx_reports_type ON reports (report_type, report_date DESC)
-  `,
-  // User-defined custom metric types (extracted from user_settings JSONB)
-  custom_metrics: `
-    CREATE TABLE IF NOT EXISTS custom_metrics (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      name            VARCHAR(100) NOT NULL UNIQUE,
-      unit            VARCHAR(30) NOT NULL,
-      description     TEXT,
-      min_value       DOUBLE PRECISION,
-      max_value       DOUBLE PRECISION,
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      updated_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-  custom_metrics_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_custom_metrics_name ON custom_metrics (name)
-  `,
-
-  // Activity type definitions (built-in + custom)
-  activity_type_definitions: `
-    CREATE TABLE IF NOT EXISTS activity_type_definitions (
-      name              VARCHAR(100) PRIMARY KEY,
-      display_name      VARCHAR(255) NOT NULL,
-      display_category  VARCHAR(50) NOT NULL DEFAULT 'other',
-      color             VARCHAR(7) NOT NULL DEFAULT '#6b7280',
-      icon              TEXT,
-      aliases           TEXT[] NOT NULL DEFAULT '{}',
-      health_connect_record_type VARCHAR(100),
-      health_connect_exercise_type INTEGER,
-      is_builtin        BOOLEAN NOT NULL DEFAULT false,
-      show_on_timeline  BOOLEAN NOT NULL DEFAULT true,
-      data_schema       JSONB,
-      parent_type       VARCHAR(100) REFERENCES activity_type_definitions(name) ON UPDATE CASCADE,
-      created_at        TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      updated_at        TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-  activity_type_definitions_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_atd_aliases ON activity_type_definitions USING GIN (aliases);
-    CREATE INDEX IF NOT EXISTS idx_atd_parent ON activity_type_definitions (parent_type)
-  `,
-  activity_type_definitions_seed: `
-    INSERT INTO activity_type_definitions (name, display_name, display_category, color, is_builtin, health_connect_record_type, health_connect_exercise_type) VALUES
-      ('sleep', 'Sleep', 'sleep_rest', '#3b82f6', true, 'SleepSessionRecord', NULL),
-      ('exercise', 'Exercise', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', NULL),
-      ('meditation', 'Meditation', 'meditation', '#a855f7', true, NULL, NULL),
-      ('nap', 'Nap', 'sleep_rest', '#60a5fa', true, 'SleepSessionRecord', NULL),
-      ('rest', 'Rest', 'sleep_rest', '#86efac', true, NULL, NULL),
-      ('calendar_event', 'Calendar Event', 'other', '#f59e0b', true, NULL, NULL),
-      ('back_extension', 'Back Extension', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 1),
-      ('badminton', 'Badminton', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 2),
-      ('barbell_shoulder_press', 'Barbell Shoulder Press', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 3),
-      ('baseball', 'Baseball', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 4),
-      ('basketball', 'Basketball', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 5),
-      ('bench_press', 'Bench Press', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 6),
-      ('bench_sit_up', 'Bench Sit Up', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 7),
-      ('biking', 'Biking', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 8),
-      ('biking_stationary', 'Biking Stationary', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 9),
-      ('boot_camp', 'Boot Camp', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 10),
-      ('boxing', 'Boxing', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 11),
-      ('burpee', 'Burpee', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 12),
-      ('calisthenics', 'Calisthenics', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 13),
-      ('cricket', 'Cricket', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 14),
-      ('crunch', 'Crunch', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 15),
-      ('dancing', 'Dancing', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 16),
-      ('deadlift', 'Deadlift', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 17),
-      ('dumbbell_curl_left_arm', 'Dumbbell Curl Left Arm', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 18),
-      ('dumbbell_curl_right_arm', 'Dumbbell Curl Right Arm', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 19),
-      ('dumbbell_front_raise', 'Dumbbell Front Raise', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 20),
-      ('dumbbell_lateral_raise', 'Dumbbell Lateral Raise', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 21),
-      ('dumbbell_triceps_extension_left_arm', 'Dumbbell Triceps Extension Left Arm', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 22),
-      ('dumbbell_triceps_extension_right_arm', 'Dumbbell Triceps Extension Right Arm', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 23),
-      ('dumbbell_triceps_extension_two_arm', 'Dumbbell Triceps Extension Two Arm', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 24),
-      ('elliptical', 'Elliptical', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 25),
-      ('exercise_class', 'Exercise Class', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 26),
-      ('fencing', 'Fencing', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 27),
-      ('football_american', 'Football American', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 28),
-      ('football_australian', 'Football Australian', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 29),
-      ('forward_twist', 'Forward Twist', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 30),
-      ('frisbee_disc', 'Frisbee Disc', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 31),
-      ('golf', 'Golf', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 32),
-      ('guided_breathing', 'Guided Breathing', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 33),
-      ('gymnastics', 'Gymnastics', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 34),
-      ('handball', 'Handball', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 35),
-      ('high_intensity_interval_training', 'High Intensity Interval Training', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 36),
-      ('hiking', 'Hiking', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 37),
-      ('ice_hockey', 'Ice Hockey', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 38),
-      ('ice_skating', 'Ice Skating', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 39),
-      ('jumping_jack', 'Jumping Jack', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 40),
-      ('jump_rope', 'Jump Rope', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 41),
-      ('lat_pull_down', 'Lat Pull Down', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 42),
-      ('lunge', 'Lunge', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 43),
-      ('martial_arts', 'Martial Arts', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 44),
-      ('other_workout', 'Other Workout', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 0),
-      ('paddling', 'Paddling', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 46),
-      ('paragliding', 'Paragliding', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 47),
-      ('pilates', 'Pilates', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 48),
-      ('plank', 'Plank', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 49),
-      ('racquetball', 'Racquetball', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 50),
-      ('rock_climbing', 'Rock Climbing', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 51),
-      ('roller_hockey', 'Roller Hockey', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 52),
-      ('rowing', 'Rowing', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 53),
-      ('rowing_machine', 'Rowing Machine', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 54),
-      ('rugby', 'Rugby', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 55),
-      ('running', 'Running', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 56),
-      ('running_treadmill', 'Running Treadmill', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 57),
-      ('sailing', 'Sailing', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 58),
-      ('scuba_diving', 'Scuba Diving', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 59),
-      ('skating', 'Skating', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 60),
-      ('skiing', 'Skiing', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 61),
-      ('snowboarding', 'Snowboarding', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 62),
-      ('snowshoeing', 'Snowshoeing', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 63),
-      ('soccer', 'Soccer', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 64),
-      ('softball', 'Softball', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 65),
-      ('squash', 'Squash', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 66),
-      ('squat', 'Squat', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 67),
-      ('stair_climbing', 'Stair Climbing', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 68),
-      ('stair_climbing_machine', 'Stair Climbing Machine', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 69),
-      ('strength_training', 'Strength Training', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 70),
-      ('stretching', 'Stretching', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 71),
-      ('surfing', 'Surfing', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 72),
-      ('swimming_open_water', 'Swimming Open Water', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 73),
-      ('swimming_pool', 'Swimming Pool', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 74),
-      ('table_tennis', 'Table Tennis', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 75),
-      ('tennis', 'Tennis', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 76),
-      ('upper_twist', 'Upper Twist', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 77),
-      ('volleyball', 'Volleyball', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 78),
-      ('walking', 'Walking', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 79),
-      ('water_polo', 'Water Polo', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 80),
-      ('weightlifting', 'Weightlifting', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 81),
-      ('wheelchair', 'Wheelchair', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 82),
-      ('yoga', 'Yoga', 'exercise', '#22c55e', true, 'ExerciseSessionRecord', 83)
-    ON CONFLICT (name) DO NOTHING;
-    -- Seed the built-in hierarchy: exercise subtypes → 'exercise', nap/rest → 'sleep'.
-    UPDATE activity_type_definitions
-       SET parent_type = 'exercise'
-     WHERE is_builtin = true
-       AND display_category = 'exercise'
-       AND name != 'exercise'
-       AND parent_type IS NULL;
-    UPDATE activity_type_definitions
-       SET parent_type = 'sleep'
-     WHERE is_builtin = true
-       AND name IN ('nap', 'rest')
-       AND parent_type IS NULL;
-    -- music_scrobble type: point-in-time activity for a Last.fm scrobble.
-    -- show_on_timeline=false because scrobbles are rendered on the dedicated
-    -- music staff track, not the main activity lane.
-    INSERT INTO activity_type_definitions (name, display_name, display_category, color, icon, is_builtin, show_on_timeline, data_schema) VALUES
-      ('music_scrobble', 'Music Scrobble', 'other', '#ec4899', '🎵', true, false, $json$
-        {"fields":[
-          {"name":"artist","type":"string","required":true,"show_in_summary":true,"is_categorical":true},
-          {"name":"track","type":"string","required":true,"show_in_summary":true},
-          {"name":"album","type":"string","required":false}
-        ]}
-      $json$::jsonb)
-    ON CONFLICT (name) DO NOTHING;
-    -- screentime type: time-span activity for merged categorized screen time.
-    -- show_on_timeline=false because screentime is rendered on its own
-    -- dedicated track, not the main activity lane. category_path is stored
-    -- as joined string so chart breakdowns by the field just work.
-    INSERT INTO activity_type_definitions (name, display_name, display_category, color, icon, is_builtin, show_on_timeline, data_schema) VALUES
-      ('screentime', 'Screen Time', 'productivity', '#64748b', '💻', true, false, $json$
-        {"fields":[
-          {"name":"category_path","type":"string","required":true,"show_in_summary":true,"is_categorical":true},
-          {"name":"score","type":"number","required":false,"show_in_summary":true,"unit":""}
-        ]}
-      $json$::jsonb)
-    ON CONFLICT (name) DO NOTHING;
-    -- location_visit type: time-span activity for a visit to an opted-in
-    -- named location. Rendered on the dedicated location track (show_on_timeline=false).
-    INSERT INTO activity_type_definitions (name, display_name, display_category, color, icon, is_builtin, show_on_timeline, data_schema) VALUES
-      ('location_visit', 'Location Visit', 'travel', '#0ea5e9', '📍', true, false, $json$
-        {"fields":[
-          {"name":"location_name","type":"string","required":true,"show_in_summary":true,"is_categorical":true},
-          {"name":"lat","type":"number","required":false},
-          {"name":"lon","type":"number","required":false}
-        ]}
-      $json$::jsonb)
-    ON CONFLICT (name) DO NOTHING
-  `,
-
-  // Deduction rules — automatically create activities from data conditions
-  deduction_rules: `
-    CREATE TABLE IF NOT EXISTS deduction_rules (
-      id                    UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      name                  VARCHAR(255) NOT NULL,
-      enabled               BOOLEAN NOT NULL DEFAULT true,
-      priority              INTEGER NOT NULL DEFAULT 0,
-      conditions            JSONB NOT NULL,
-      output_activity_type  VARCHAR(100) NOT NULL REFERENCES activity_type_definitions(name) ON UPDATE CASCADE,
-      output_title          VARCHAR(255),
-      merge_gap_seconds     INTEGER,
-      created_at            TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      updated_at            TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-  deduction_rules_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_deduction_rules_enabled ON deduction_rules (enabled, priority) WHERE enabled = true
-  `,
-
-  // Deduction rule run audit log
-  deduction_rule_runs: `
-    CREATE TABLE IF NOT EXISTS deduction_rule_runs (
-      id                UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      rule_id           UUID NOT NULL REFERENCES deduction_rules(id) ON DELETE CASCADE,
-      evaluated_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      window_start      TIMESTAMPTZ NOT NULL,
-      window_end        TIMESTAMPTZ NOT NULL,
-      activities_created INTEGER NOT NULL DEFAULT 0,
-      duration_ms       INTEGER
-    )
-  `,
-  deduction_rule_runs_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_deduction_rule_runs_rule ON deduction_rule_runs (rule_id, evaluated_at DESC)
-  `,
-
-  // User-defined goals for tracking metrics (extracted from user_settings JSONB)
-  goals: `
-    CREATE TABLE IF NOT EXISTS goals (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      goal_type       VARCHAR(10) NOT NULL DEFAULT 'metric',
-      metric          VARCHAR(50),
-      min_value       DOUBLE PRECISION,
-      max_value       DOUBLE PRECISION,
-      time_window     VARCHAR(10) DEFAULT '7d',
-      source_type     VARCHAR(30),
-      pattern         TEXT,
-      half_life_days  INTEGER,
-      display_period  VARCHAR(10),
-      aggregation     VARCHAR(10),
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      updated_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-  goals_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_goals_metric ON goals (metric)
-  `,
-
-  // Screentime category rules for categorizing productivity records
-
-  screentime_categories: `
-    CREATE TABLE IF NOT EXISTS screentime_categories (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      name            TEXT[] NOT NULL,
-      rule_type       VARCHAR(20) NOT NULL DEFAULT 'none',
-      rule_regex      TEXT,
-      ignore_case     BOOLEAN DEFAULT TRUE,
-      color           VARCHAR(20),
-      score           SMALLINT,
-      exclude_from_screentime BOOLEAN DEFAULT FALSE,
-      sort_order      INTEGER DEFAULT 0,
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      updated_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-  screentime_categories_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_screentime_categories_name ON screentime_categories USING GIN (name)
-  `,
-
-  // Sync state tracking for incremental data pulls
-  sync_state: `
-    CREATE TABLE IF NOT EXISTS sync_state (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      provider        VARCHAR(50) NOT NULL,
-      data_type       VARCHAR(100) NOT NULL,
-      last_sync_time  TIMESTAMPTZ,
-      sync_start_date DATE,
-      status          VARCHAR(20) DEFAULT 'idle',
-      error_message   TEXT,
-      retry_after     TIMESTAMPTZ,
-      updated_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      CONSTRAINT unique_sync_state UNIQUE (provider, data_type)
-    )
-  `,
-
-  // Legacy table — kept for migration only, no new data written
-  tag_definitions: `
-    CREATE TABLE IF NOT EXISTS tag_definitions (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      name            VARCHAR(100) NOT NULL,
-      icon            TEXT,
-      show_on_timeline BOOLEAN NOT NULL DEFAULT true,
-      aliases         TEXT[] NOT NULL DEFAULT '{}',
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      updated_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-  tag_definitions_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_tag_definitions_aliases ON tag_definitions USING GIN (aliases);
-    CREATE UNIQUE INDEX IF NOT EXISTS idx_tag_definitions_name ON tag_definitions (lower(name))
-  `,
-
-  // Legacy table — kept for migration only, no new data written
-  tags: `
-    CREATE TABLE IF NOT EXISTS tags (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      source          VARCHAR(50) NOT NULL,
-      external_id     VARCHAR(255),
-      tag             VARCHAR(100) NOT NULL,
-      tag_key         VARCHAR(255),
-      tag_definition_id UUID REFERENCES tag_definitions(id),
-      start_time      TIMESTAMPTZ NOT NULL,
-      end_time        TIMESTAMPTZ,
-      deleted_at      TIMESTAMPTZ,
-      CONSTRAINT unique_tag UNIQUE (source, external_id)
-    )
-  `,
-  tags_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_tags_time ON tags (start_time DESC);
-    CREATE INDEX IF NOT EXISTS idx_tags_tag_time ON tags (tag, start_time DESC);
-    CREATE INDEX IF NOT EXISTS idx_tags_tag_key ON tags (tag_key) WHERE tag_key IS NOT NULL;
-    CREATE INDEX IF NOT EXISTS idx_tags_not_deleted ON tags (start_time DESC) WHERE deleted_at IS NULL;
-    CREATE INDEX IF NOT EXISTS idx_tags_definition_id ON tags (tag_definition_id) WHERE tag_definition_id IS NOT NULL
-  `,
-
-  // Normalized time-series metrics for fast charting queries
-  time_series: `
-    CREATE TABLE IF NOT EXISTS time_series (
-      time            TIMESTAMPTZ NOT NULL,
-      metric          VARCHAR(50) NOT NULL,
-      value           DOUBLE PRECISION NOT NULL,
-      unit            VARCHAR(20) NOT NULL,
-      source          VARCHAR(50) NOT NULL,
-      deleted_at      TIMESTAMPTZ,
-      PRIMARY KEY (time, metric, source)
-    )
-  `,
-  time_series_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_time_series_metric_time ON time_series (metric, time DESC);
-    CREATE INDEX IF NOT EXISTS idx_time_series_metric_source_time ON time_series (metric, source, time DESC)
-  `,
-
-  // Uploaded icon images (stored as binary blobs)
-  uploaded_icons: `
-    CREATE TABLE IF NOT EXISTS uploaded_icons (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      content_type    VARCHAR(50) NOT NULL,
-      data            BYTEA NOT NULL,
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-
-  // User settings (HR zones, birth date, etc.)
-  user_settings: `
-    CREATE TABLE IF NOT EXISTS user_settings (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      settings        JSONB NOT NULL DEFAULT '{}',
-      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      updated_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
-    )
-  `,
-
-  // Audit log for user-specific events (sync, auth, settings changes, etc.)
-  audit_log: `
-    CREATE TABLE IF NOT EXISTS audit_log (
-      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-      timestamp       TIMESTAMPTZ NOT NULL DEFAULT NOW(),
-      level           VARCHAR(10) NOT NULL DEFAULT 'info',
-      category        VARCHAR(20) NOT NULL,
-      message         TEXT NOT NULL,
-      details         JSONB
-    )
-  `,
-
-  audit_log_indexes: `
-    CREATE INDEX IF NOT EXISTS idx_audit_log_timestamp ON audit_log (timestamp DESC);
-    CREATE INDEX IF NOT EXISTS idx_audit_log_category ON audit_log (category);
-    CREATE INDEX IF NOT EXISTS idx_audit_log_level ON audit_log (level)
-  `,
+  ...activitiesTables,
+  ...locationsTables,
+  ...mealsTables,
+  ...metricsTables,
+  ...productivityTables,
+  ...reportsTables,
+  ...systemTables,
 }
 
 /**
  * Order in which tables should be created (respecting dependencies).
+ *
+ * Key constraints:
+ *  - activity_type_definitions before activities (FK)
+ *  - reports before report_entries (FK)
+ *  - meals + food_items before meal_food_items (FK)
+ *  - deduction_rules before deduction_rule_runs (FK)
+ *  - tag_definitions before tags (FK)
  */
 export const tableCreationOrder = [
   'raw_records',
