+++ apps/backend/src/schema/meals.ts
@@ -1,0 +1,209 @@
+/**
+ * Meal-domain table SQL: meals, meal log completion, food items, meal-food
+ * junction. Food items are normalized; meal_food_items snapshots nutrient
+ * values at insertion time so changes to the canonical food don't
+ * retroactively alter past meals.
+ */
+export const mealsTables: Record<string, string> = {
+  // Meal/nutrition data
+  meals: `
+    CREATE TABLE IF NOT EXISTS meals (
+      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
+      source          VARCHAR(50) NOT NULL DEFAULT 'manual',
+      meal_type       VARCHAR(50),
+      name            VARCHAR(255),
+      time            TIMESTAMPTZ NOT NULL,
+      calories        DOUBLE PRECISION,
+      protein         DOUBLE PRECISION,
+      carbs           DOUBLE PRECISION,
+      fat             DOUBLE PRECISION,
+      fiber           DOUBLE PRECISION,
+      food_items      JSONB,
+      micros          JSONB,
+      notes           TEXT,
+      sensitivities   TEXT[],
+      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
+    )
+  `,
+
+  // Meal log completion flag per day
+  meal_log_completed: `
+    CREATE TABLE IF NOT EXISTS meal_log_completed (
+      date            DATE PRIMARY KEY,
+      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
+    )
+  `,
+
+  // Canonical food item library (normalized from JSONB)
+  food_items: `
+    CREATE TABLE IF NOT EXISTS food_items (
+      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
+      name            VARCHAR(255) NOT NULL,
+      name_lower      VARCHAR(255) NOT NULL,
+      source          VARCHAR(50) NOT NULL DEFAULT 'manual',
+      default_quantity DOUBLE PRECISION,
+      default_unit    VARCHAR(100),
+      calories        DOUBLE PRECISION,
+      protein         DOUBLE PRECISION,
+      carbs           DOUBLE PRECISION,
+      fat             DOUBLE PRECISION,
+      fiber           DOUBLE PRECISION,
+      alcohol         DOUBLE PRECISION,
+      caffeine        DOUBLE PRECISION,
+      water           DOUBLE PRECISION,
+      net_carbs       DOUBLE PRECISION,
+      starch          DOUBLE PRECISION,
+      sugars          DOUBLE PRECISION,
+      added_sugars    DOUBLE PRECISION,
+      cholesterol     DOUBLE PRECISION,
+      saturated_fat   DOUBLE PRECISION,
+      monounsaturated_fat DOUBLE PRECISION,
+      polyunsaturated_fat DOUBLE PRECISION,
+      trans_fat       DOUBLE PRECISION,
+      omega_3         DOUBLE PRECISION,
+      omega_6         DOUBLE PRECISION,
+      ala             DOUBLE PRECISION,
+      dha             DOUBLE PRECISION,
+      epa             DOUBLE PRECISION,
+      dpa             DOUBLE PRECISION,
+      aa              DOUBLE PRECISION,
+      la              DOUBLE PRECISION,
+      vitamin_a       DOUBLE PRECISION,
+      retinol         DOUBLE PRECISION,
+      beta_carotene   DOUBLE PRECISION,
+      vitamin_c       DOUBLE PRECISION,
+      vitamin_d       DOUBLE PRECISION,
+      vitamin_e       DOUBLE PRECISION,
+      vitamin_k       DOUBLE PRECISION,
+      b1_thiamine     DOUBLE PRECISION,
+      b2_riboflavin   DOUBLE PRECISION,
+      b3_niacin       DOUBLE PRECISION,
+      b5_pantothenic_acid DOUBLE PRECISION,
+      b6_pyridoxine   DOUBLE PRECISION,
+      b12_cobalamin   DOUBLE PRECISION,
+      folate          DOUBLE PRECISION,
+      calcium         DOUBLE PRECISION,
+      copper          DOUBLE PRECISION,
+      iron            DOUBLE PRECISION,
+      magnesium       DOUBLE PRECISION,
+      manganese       DOUBLE PRECISION,
+      phosphorus      DOUBLE PRECISION,
+      potassium       DOUBLE PRECISION,
+      selenium        DOUBLE PRECISION,
+      sodium          DOUBLE PRECISION,
+      zinc            DOUBLE PRECISION,
+      iodine          DOUBLE PRECISION,
+      cystine         DOUBLE PRECISION,
+      histidine       DOUBLE PRECISION,
+      isoleucine      DOUBLE PRECISION,
+      leucine         DOUBLE PRECISION,
+      lysine          DOUBLE PRECISION,
+      methionine      DOUBLE PRECISION,
+      phenylalanine   DOUBLE PRECISION,
+      threonine       DOUBLE PRECISION,
+      tryptophan      DOUBLE PRECISION,
+      tyrosine        DOUBLE PRECISION,
+      valine          DOUBLE PRECISION,
+      oxalate         DOUBLE PRECISION,
+      phytate         DOUBLE PRECISION,
+      ash             DOUBLE PRECISION,
+      salt            DOUBLE PRECISION,
+      icon            TEXT,
+      created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
+      updated_at      TIMESTAMPTZ NOT NULL DEFAULT NOW(),
+      CONSTRAINT unique_food_item_name UNIQUE (name_lower)
+    )
+  `,
+
+  food_items_indexes: `
+    CREATE INDEX IF NOT EXISTS idx_food_items_name_lower ON food_items (name_lower)
+  `,
+
+  // Junction: meals <-> food items (snapshot of nutrients at insertion time)
+  meal_food_items: `
+    CREATE TABLE IF NOT EXISTS meal_food_items (
+      id              UUID PRIMARY KEY DEFAULT gen_random_uuid(),
+      meal_id         UUID NOT NULL REFERENCES meals(id) ON DELETE CASCADE,
+      food_item_id    UUID NOT NULL REFERENCES food_items(id),
+      quantity        DOUBLE PRECISION,
+      unit            VARCHAR(100),
+      sort_order      INTEGER NOT NULL DEFAULT 0,
+      calories        DOUBLE PRECISION,
+      protein         DOUBLE PRECISION,
+      carbs           DOUBLE PRECISION,
+      fat             DOUBLE PRECISION,
+      fiber           DOUBLE PRECISION,
+      alcohol         DOUBLE PRECISION,
+      caffeine        DOUBLE PRECISION,
+      water           DOUBLE PRECISION,
+      net_carbs       DOUBLE PRECISION,
+      starch          DOUBLE PRECISION,
+      sugars          DOUBLE PRECISION,
+      added_sugars    DOUBLE PRECISION,
+      cholesterol     DOUBLE PRECISION,
+      saturated_fat   DOUBLE PRECISION,
+      monounsaturated_fat DOUBLE PRECISION,
+      polyunsaturated_fat DOUBLE PRECISION,
+      trans_fat       DOUBLE PRECISION,
+      omega_3         DOUBLE PRECISION,
+      omega_6         DOUBLE PRECISION,
+      ala             DOUBLE PRECISION,
+      dha             DOUBLE PRECISION,
+      epa             DOUBLE PRECISION,
+      dpa             DOUBLE PRECISION,
+      aa              DOUBLE PRECISION,
+      la              DOUBLE PRECISION,
+      vitamin_a       DOUBLE PRECISION,
+      retinol         DOUBLE PRECISION,
+      beta_carotene   DOUBLE PRECISION,
+      vitamin_c       DOUBLE PRECISION,
+      vitamin_d       DOUBLE PRECISION,
+      vitamin_e       DOUBLE PRECISION,
+      vitamin_k       DOUBLE PRECISION,
+      b1_thiamine     DOUBLE PRECISION,
+      b2_riboflavin   DOUBLE PRECISION,
+      b3_niacin       DOUBLE PRECISION,
+      b5_pantothenic_acid DOUBLE PRECISION,
+      b6_pyridoxine   DOUBLE PRECISION,
+      b12_cobalamin   DOUBLE PRECISION,
+      folate          DOUBLE PRECISION,
+      calcium         DOUBLE PRECISION,
+      copper          DOUBLE PRECISION,
+      iron            DOUBLE PRECISION,
+      magnesium       DOUBLE PRECISION,
+      manganese       DOUBLE PRECISION,
+      phosphorus      DOUBLE PRECISION,
+      potassium       DOUBLE PRECISION,
+      selenium        DOUBLE PRECISION,
+      sodium          DOUBLE PRECISION,
+      zinc            DOUBLE PRECISION,
+      iodine          DOUBLE PRECISION,
+      cystine         DOUBLE PRECISION,
+      histidine       DOUBLE PRECISION,
+      isoleucine      DOUBLE PRECISION,
+      leucine         DOUBLE PRECISION,
+      lysine          DOUBLE PRECISION,
+      methionine      DOUBLE PRECISION,
+      phenylalanine   DOUBLE PRECISION,
+      threonine       DOUBLE PRECISION,
+      tryptophan      DOUBLE PRECISION,
+      tyrosine        DOUBLE PRECISION,
+      valine          DOUBLE PRECISION,
+      oxalate         DOUBLE PRECISION,
+      phytate         DOUBLE PRECISION,
+      ash             DOUBLE PRECISION,
+      salt            DOUBLE PRECISION
+    )
+  `,
+
+  meal_food_items_indexes: `
+    CREATE INDEX IF NOT EXISTS idx_meal_food_items_meal ON meal_food_items (meal_id);
+    CREATE INDEX IF NOT EXISTS idx_meal_food_items_food ON meal_food_items (food_item_id)
+  `,
+
+  meals_indexes: `
+    CREATE INDEX IF NOT EXISTS idx_meals_time ON meals (time DESC);
+    CREATE INDEX IF NOT EXISTS idx_meals_type_time ON meals (meal_type, time DESC);
+    CREATE INDEX IF NOT EXISTS idx_meals_source ON meals (source, time DESC)
+  `,
+}
